+++ README.md
@@ -29,7 +29,27 @@ of endmill damage.
 ![Example path](/images/longneck.png)
 
 ## How to use this library
-See `demo.py` file.
+### Basic calculation of a internal pocket:
+```
+$ ./examples/example_basic.py ./test_cases/stegasorus.dxf
+```
+### Calculation of a internal pocket starting from pre-drilled hole:
+```
+./examples/example_with_predrill.py ./test_cases/triceratops.dxf
+```
+### Calculation of an outside clearing operation where the edge of the stock to be cut is encoded as the outside polygon in the .dxf file:
+```
+examples/example_outer.py ./test_cases/hextest.dxf
+```
+### Calculation of an outside clearing operation where the edge of the stock to be cut is calculated as a simple offset from the outer edge of the .dxf part file.
+```
+examples/example_refine.py ./test_cases/thrash.dxf
+```
+
+## Other useful libraries in this repo
+The VoronoiCenters class in hsm_nibble/voronoi_centers.py is a wrapper around the pyvoronoi package.
+pyvoronoi only deals with straight lines and points. Not arcs.
+The VoronoiCenters class prunes the voroni diagram and tries to do same things when simplifying a voronoi diagram created from geometry containing arc and circles.
 
 ## Projects using this library
 - [DerpCAM](https://github.com/kfoltman/DerpCAM) A 2.5D CAM aimed at hobby CNC machine users.
