+++ hsm_nibble/helpers.py
