+++ pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "hsm_nibble"
+version = "0.1.0"
+authors = [
+  { name="dunk", email="mrdunk@gmail.com" },
+]
+description = "A library for generating HSM tool paths for CNC pocket milling."
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: GNU Affero General Public License v3",
+    "Operating System :: OS Independent",
+    "Topic :: Scientific/Engineering"
+]
+
+[project.urls]
+"Homepage" = "https://github.com/mrdunk/HSM_nibbler"
