+++ tests/test_voronoi.py
@@ -8,7 +8,7 @@ from shapely.geometry import LineString, Polygon  # type: ignore
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-import voronoi_centers
+import hsm_nibble.voronoi_centers as voronoi_centers
 
 class TestBoilerplate(unittest.TestCase):
     def setUp(self):
